Refactor Account and Host into separate modules
