+++ Host.js
@@ -1,0 +1,11 @@
+import { img } from 'https://unpkg.com/ez-html-elements'
+
+export default (host) => {
+  return Object.freeze({
+    faviconHtml: () =>
+      img(['inline'], {
+        src: `https://${host}/favicon.ico`,
+        alt: host
+      })
+  })
+}
